Move callbacks to Events. Start separating logic more.
