+++ src/libs/Network/index.js
@@ -1,143 +1,16 @@
 import _ from 'underscore';
 import lodashGet from 'lodash/get';
-import Onyx from 'react-native-onyx';
 import HttpUtils from '../HttpUtils';
-import ONYXKEYS from '../../ONYXKEYS';
 import * as ActiveClientManager from '../ActiveClientManager';
 import CONST from '../../CONST';
-import createCallback from '../createCallback';
 import * as NetworkRequestQueue from '../actions/NetworkRequestQueue';
 import * as NetworkStore from './NetworkStore';
-import enhanceParameters from './enhanceParameters';
-
-let isOffline = false;
-let persistedRequestsQueueRunning = false;
+import * as NetworkEvents from './NetworkEvents';
+import * as PersistedRequestsQueue from './PersistedRequestsQueue';
+import processRequest from './processRequest';
 
 // Queue for network requests so we don't lose actions done by the user while offline
 let networkRequestQueue = [];
-
-// These handlers must be registered so we can process the request, response, and errors returned from the queue.
-// The first argument passed will be the queuedRequest object and the second will be either the parameters, response, or error.
-const [onRequest, registerRequestHandler] = createCallback();
-const [onResponse, registerResponseHandler] = createCallback();
-const [onError, registerErrorHandler] = createCallback();
-const [onRequestSkipped, registerRequestSkippedHandler] = createCallback();
-const [getLogger, registerLogHandler] = createCallback();
-const [recheckConnectivity, registerConnectionCheckCallback] = createCallback();
-
-/**
- * @param {Object} request
- * @param {String} request.command
- * @param {Object} request.data
- * @param {String} request.type
- * @param {Boolean} request.shouldUseSecure
- * @returns {Promise}
- */
-function processRequest(request) {
-    const finalParameters = enhanceParameters(request.command, request.data);
-
-    // If request is still in processing after this time, we might be offline
-    const timerId = setTimeout(recheckConnectivity, CONST.NETWORK.MAX_PENDING_TIME_MS);
-
-    onRequest(request, finalParameters);
-    return HttpUtils.xhr(request.command, finalParameters, request.type, request.shouldUseSecure)
-        .finally(() => clearTimeout(timerId));
-}
-
-/**
- * This method will get any persisted requests and fire them off in parallel to retry them.
- * If we get any jsonCode besides 407 the request is a success. It doesn't make sense to
- * continually retry things that have returned a response. However, we can retry any requests
- * with known networking errors like "Failed to fetch".
- *
- * @returns {Promise}
- */
-function processPersistedRequestsQueue() {
-    const persistedRequests = NetworkRequestQueue.getPersistedRequests();
-
-    // This sanity check is also a recursion exit point
-    if (isOffline || _.isEmpty(persistedRequests)) {
-        return Promise.resolve();
-    }
-
-    const tasks = _.map(persistedRequests, request => processRequest(request)
-        .then((response) => {
-            if (response.jsonCode === CONST.JSON_CODE.NOT_AUTHENTICATED) {
-                getLogger().info('Persisted optimistic request needs authentication');
-            } else {
-                getLogger().info('Persisted optimistic request returned a valid jsonCode. Not retrying.');
-            }
-            onResponse(request, response);
-            NetworkRequestQueue.removeRetryableRequest(request);
-        })
-        .catch((error) => {
-            // If we are catching a known network error like "Failed to fetch" allow this request to be retried if we have retries left
-            if (error.message === CONST.ERROR.FAILED_TO_FETCH) {
-                const retryCount = NetworkRequestQueue.incrementRetries(request);
-                getLogger().info('Persisted request failed', false, {retryCount, command: request.command, error: error.message});
-                if (retryCount >= CONST.NETWORK.MAX_REQUEST_RETRIES) {
-                    getLogger().info('Request failed too many times removing from storage', false, {retryCount, command: request.command, error: error.message});
-                    NetworkRequestQueue.removeRetryableRequest(request);
-                }
-            } else if (error.name === CONST.ERROR.REQUEST_CANCELLED) {
-                getLogger().info('Persisted request was cancelled. Not retrying.');
-                onError(request);
-                NetworkRequestQueue.removeRetryableRequest(request);
-            } else {
-                getLogger().alert(`${CONST.ERROR.ENSURE_BUGBOT} unknown error while retrying persisted request. Not retrying.`, {
-                    command: request.command,
-                    error: error.message,
-                });
-                NetworkRequestQueue.removeRetryableRequest(request);
-            }
-        }));
-
-    // Do a recursive call in case the queue is not empty after processing the current batch
-    return Promise.all(tasks)
-        .then(processPersistedRequestsQueue);
-}
-
-function flushPersistedRequestsQueue() {
-    if (persistedRequestsQueueRunning) {
-        return;
-    }
-
-    // NETWORK_REQUEST_QUEUE is shared across clients, thus every client/tab will have a copy
-    // It is very important to only process the queue from leader client otherwise requests will be duplicated.
-    if (!ActiveClientManager.isClientTheLeader()) {
-        return;
-    }
-
-    persistedRequestsQueueRunning = true;
-
-    // Ensure persistedRequests are read from storage before proceeding with the queue
-    const connectionId = Onyx.connect({
-        key: ONYXKEYS.NETWORK_REQUEST_QUEUE,
-        callback: () => {
-            Onyx.disconnect(connectionId);
-            processPersistedRequestsQueue()
-                .finally(() => persistedRequestsQueueRunning = false);
-        },
-    });
-}
-
-// We subscribe to the online/offline status of the network to determine when we should fire off API calls
-// vs queueing them for later.
-Onyx.connect({
-    key: ONYXKEYS.NETWORK,
-    callback: (network) => {
-        if (!network) {
-            return;
-        }
-
-        // Client becomes online, process the queue.
-        if (isOffline && !network.isOffline) {
-            flushPersistedRequestsQueue();
-        }
-
-        isOffline = network.isOffline;
-    },
-});
 
 /**
  * Checks to see if a request can be made.
@@ -151,7 +24,7 @@ Onyx.connect({
  */
 function canMakeRequest(request) {
     if (!NetworkStore.isReady()) {
-        onRequestSkipped({command: request.command, type: request.type});
+        NetworkEvents.onRequestSkipped({command: request.command, type: request.type});
         return false;
     }
 
@@ -198,7 +71,7 @@ function canRetryRequest(request) {
  */
 function processNetworkRequestQueue() {
     // NetInfo tells us whether the app is offline
-    if (isOffline) {
+    if (NetworkStore.getIsOffline()) {
         if (!networkRequestQueue.length) {
             return;
         }
@@ -243,15 +116,15 @@ function processNetworkRequestQueue() {
         }
 
         processRequest(queuedRequest)
-            .then(response => onResponse(queuedRequest, response))
+            .then(response => NetworkEvents.onResponse(queuedRequest, response))
             .catch((error) => {
                 // Cancelled requests should not be retried
                 if (error.name === CONST.ERROR.REQUEST_CANCELLED) {
-                    onError(queuedRequest, error);
+                    NetworkEvents.onError(queuedRequest, error);
                     return;
                 }
 
-                recheckConnectivity();
+                NetworkEvents.triggerRecheckNeeded();
 
                 // Retry and request that returns a "Failed to fetch" error. Very common if a user is offline or experiencing an unlikely scenario.
                 if (error.message === CONST.ERROR.FAILED_TO_FETCH) {
@@ -259,7 +132,7 @@ function processNetworkRequestQueue() {
                     const shouldRetry = lodashGet(queuedRequest, 'data.shouldRetry');
                     if (shouldRetry) {
                         const retryCount = NetworkRequestQueue.incrementRetries(queuedRequest);
-                        getLogger().info('A retryable request failed', false, {
+                        NetworkEvents.getLogger().info('A retryable request failed', false, {
                             retryCount,
                             command: queuedRequest.command,
                             error: error.message,
@@ -270,12 +143,12 @@ function processNetworkRequestQueue() {
                             return;
                         }
 
-                        getLogger().info('Request was retried too many times with no success. No more retries left');
+                        NetworkEvents.getLogger().info('Request was retried too many times with no success. No more retries left');
                     }
 
-                    onError(queuedRequest, error);
+                    NetworkEvents.onError(queuedRequest, error);
                 } else {
-                    getLogger().alert(`${CONST.ERROR.ENSURE_BUGBOT} unknown error caught while processing request`, {
+                    NetworkEvents.getLogger().alert(`${CONST.ERROR.ENSURE_BUGBOT} unknown error caught while processing request`, {
                         command: queuedRequest.command,
                         error: error.message,
                     });
@@ -294,7 +167,7 @@ function startDefaultQueue() {
 
 // Post any pending request after we launch the app
 ActiveClientManager.isReady().then(() => {
-    flushPersistedRequestsQueue();
+    PersistedRequestsQueue.flush();
     startDefaultQueue();
 });
 
@@ -358,10 +231,4 @@ function clearRequestQueue() {
 export {
     post,
     clearRequestQueue,
-    registerResponseHandler,
-    registerErrorHandler,
-    registerRequestHandler,
-    registerRequestSkippedHandler,
-    registerLogHandler,
-    registerConnectionCheckCallback,
 };
