+++ src/libs/actions/Session/index.js
@@ -13,7 +13,6 @@ import CONST from '../../../CONST';
 import Navigation from '../../Navigation/Navigation';
 import ROUTES from '../../../ROUTES';
 import * as Localize from '../../Localize';
-import * as Network from '../../Network';
 import UnreadIndicatorUpdater from '../../UnreadIndicatorUpdater';
 import Timers from '../../Timers';
 import * as Pusher from '../../Pusher/pusher';
